plugin model check test is split up to two tests
