+++ test/plugin_test_model/CMakeLists.txt
@@ -1,9 +1,42 @@
-set(testid "test_plugin_test_model")
+set(test1_id "test_plugin_test_model_1")
 
 add_test_executable(
-        ${testid} test_targets
+        ${test1_id} test_targets
         ${SharedTestFiles}
-        model_check.cpp
+        model_check_1.cpp PluginTestModelTests.cpp
+)
+
+# make test_targets list global to all tests
+set(test_targets ${test_targets} PARENT_SCOPE)
+
+set(test2_id "test_plugin_test_model_2")
+
+add_test_executable(
+        ${test2_id} test_targets
+        ${SharedTestFiles}
+        model_check_2.cpp PluginTestModelTests.cpp
+)
+
+# make test_targets list global to all tests
+set(test_targets ${test_targets} PARENT_SCOPE)
+
+set(test3_id "test_plugin_test_model_3")
+
+add_test_executable(
+        ${test3_id} test_targets
+        ${SharedTestFiles}
+        model_check_3.cpp PluginTestModelTests.cpp
+)
+
+# make test_targets list global to all tests
+set(test_targets ${test_targets} PARENT_SCOPE)
+
+set(test4_id "test_plugin_test_model_4")
+
+add_test_executable(
+        ${test4_id} test_targets
+        ${SharedTestFiles}
+        model_check_4.cpp PluginTestModelTests.cpp
 )
 
 # make test_targets list global to all tests
