+++ test/plugin_test_model/PluginTestModelTests.cpp
@@ -1,0 +1,5 @@
+#include "PluginTestModelTests.h"
+
+PluginTestModelTests::PluginTestModelTests() {
+    pluginsModelsDir = rrTestModelsDir_ / "PLUGINS";
+}
