+++ test/plugin_test_model/PluginTestModelTests.h
@@ -1,0 +1,17 @@
+#ifndef ROADRUNNER_PLUGIN_TEST_MODEL_TESTS_H
+#define ROADRUNNER_PLUGIN_TEST_MODEL_TESTS_H
+
+#include "gtest/gtest.h"
+#include <filesystem>
+#include "RoadRunnerTest.h"
+
+using std::filesystem::path;
+
+class PluginTestModelTests : public RoadRunnerTest {
+public:
+    path pluginsModelsDir;
+
+    PluginTestModelTests();
+};
+
+#endif
